+++ src/res/js/app.HttpResponse.js
@@ -1,0 +1,31 @@
+/**
+ * This file is part of the vscode-powertools distribution.
+ * Copyright (c) e.GO Digital GmbH, Aachen, Germany (https://www.e-go-digital.com/)
+ *
+ * vscode-powertools is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, version 3.
+ *
+ * vscode-powertools is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+
+function ego_on_loaded() {
+    ego_apply_highlight(
+        $('#ego-http-response')
+    );
+
+    $('#ego-save-content-btn').on('click', () => {
+        ego_post('saveContent');
+    });
+
+    $('#ego-save-request-btn').on('click', () => {
+        ego_post('saveRequest');
+    });
+}